+++ apps/web/src/features/sanity/SanityTest.tsx
@@ -1,5 +1,5 @@
 import { type PropsWithChildren, type ReactNode } from "react";
-import { cn } from "@/utilities/cn";
+import { cn } from "@/utilities/cn.ts";
 
 export type SanityTestData = {
   experiencesCount: number;